# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AppliedFilters": {
        "description": "The filters actually applied after validation and defaults (default window: 30 days\nending at the database's latest trending date). Empty country/category lists mean all.",
        "properties": {
          "categories": {
            "description": "Applied categories; empty = all categories.",
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "countries": {
            "description": "Applied country codes; empty = all countries.",
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "start_date",
          "end_date",
          "days",
          "countries",
          "categories"
        ],
        "title": "AppliedFilters",
        "type": "object"
      },
...
      "DataResponse_VideoDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VideoDetail"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[VideoDetail]",
        "type": "object"
      },
      "DataResponse_list_HistoryPoint__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/HistoryPoint"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[list[HistoryPoint]]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same value as the X-Request-ID response header.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FilteredResponse_VideoPage_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VideoPage"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[VideoPage]",
        "type": "object"
      },
...
      "HistoryPoint": {
        "description": "One trending snapshot (country x day). Views are the global lifetime count at that snapshot;\ndo not sum them across snapshots.",
        "properties": {
          "comments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "country_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Name"
          },
          "engagement_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engagement Rate"
          },
          "likes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Likes"
          },
          "trending_date": {
            "format": "date",
            "title": "Trending Date",
            "type": "string"
          },
          "views": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Views"
          }
        },
        "required": [
          "country_code",
          "country_name",
          "trending_date",
          "views",
          "likes",
          "comments",
          "engagement_rate"
        ],
        "title": "HistoryPoint",
        "type": "object"
      },
...
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "VideoDetail": {
        "description": "From app.video_details: the video's GLOBALLY latest snapshot (latest date, then highest\nviews, then country code) plus its full lifecycle summary. Not affected by filters.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelInfo"
          },
          "comments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dimension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "engagement_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engagement Rate"
          },
          "first_trending_date": {
            "format": "date",
            "title": "First Trending Date",
            "type": "string"
          },
          "last_trending_date": {
            "format": "date",
            "title": "Last Trending Date",
            "type": "string"
          },
          "latest_country_code": {
            "title": "Latest Country Code",
            "type": "string"
          },
          "latest_trending_date": {
            "format": "date",
            "title": "Latest Trending Date",
            "type": "string"
          },
          "licensed_content": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensed Content"
          },
          "likes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Likes"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "snapshot_count": {
            "title": "Snapshot Count",
            "type": "integer"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "trending_countries": {
            "items": {
              "type": "string"
            },
            "title": "Trending Countries",
            "type": "array"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "views": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Views"
          }
        },
        "required": [
          "video_id",
          "title",
          "description",
          "tags",
          "category",
          "published_at",
          "duration_sec",
          "definition",
          "dimension",
          "licensed_content",
          "thumbnail_url",
          "views",
          "likes",
          "comments",
          "engagement_rate",
          "latest_country_code",
          "latest_trending_date",
          "first_trending_date",
          "last_trending_date",
          "trending_countries",
          "snapshot_count",
          "channel"
        ],
        "title": "VideoDetail",
        "type": "object"
      },
      "VideoListItem": {
        "description": "One video. Title, channel, category, counts and engagement come from its latest snapshot\nWITHIN THE FILTERED DATASET; the date/country fields aggregate its filtered snapshots.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "channel_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Title"
          },
          "comments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "days_on_trending": {
            "description": "Distinct trending dates within the filters.",
            "title": "Days On Trending",
            "type": "integer"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "engagement_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "(likes + comments) / views; null when views are 0.",
            "title": "Engagement Rate"
          },
          "first_trending_date": {
            "format": "date",
            "title": "First Trending Date",
            "type": "string"
          },
          "last_trending_date": {
            "format": "date",
            "title": "Last Trending Date",
            "type": "string"
          },
          "likes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Likes"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "snapshot_count": {
            "description": "Filtered snapshots (country x day appearances).",
            "title": "Snapshot Count",
            "type": "integer"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "trending_countries": {
            "description": "Countries with a filtered snapshot.",
            "items": {
              "type": "string"
            },
            "title": "Trending Countries",
            "type": "array"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "views": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Global lifetime view_count at the latest filtered snapshot.",
            "title": "Views"
          }
        },
        "required": [
          "video_id",
          "title",
          "channel_id",
          "channel_title",
          "category",
          "thumbnail_url",
          "published_at",
          "duration_sec",
          "views",
          "likes",
          "comments",
          "engagement_rate",
          "trending_countries",
          "days_on_trending",
          "first_trending_date",
          "last_trending_date",
          "snapshot_count"
        ],
        "title": "VideoListItem",
        "type": "object"
      },
      "VideoPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VideoListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Videos matching the filters and search.",
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "VideoPage",
        "type": "object"
      },
      "VideoSort": {
        "enum": [
          "views",
          "likes",
          "comments",
          "engagement_rate",
          "days_on_trending",
          "first_trending_date",
          "last_trending_date",
          "published_at"
        ],
        "title": "VideoSort",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "\nRead-only analytics API over YouTube trending snapshots (PostgreSQL `app` schema) and a frozen ML model.\n\n**Metric semantics**\n* *Latest snapshot* = each video's latest snapshot **within the filtered dataset** (date range,\n  countries, categories), not its globally latest snapshot.\n* **Snapshot-level** metrics count trending appearances (country x video x day), e.g. trending volume.\n* **Video-level** metrics count each video once via its latest filtered snapshot, e.g. unique videos,\n  views, engagement rate.\n* **Country-level** metrics use each video's latest filtered snapshot in that country. Country\n  `views_of_trending_videos` are **global lifetime views** of videos that trended there; they are\n  **not additive** across countries.\n* View counts are lifetime totals: historical snapshots are **never summed** for views.\n* Period change: counts and views in **percent**; engagement rate in **percentage points**.\n\n**Authentication**: sign in via `/api/v1/auth/login` (or register). The session is an HttpOnly\ncookie; every `/api/v1` endpoint except register/login requires it. Analytics endpoints need\n`VIEW_ANALYTICS`, predictions `MAKE_PREDICTION`, member management `MANAGE_USERS`. The YouTube\ndataset is global and read-only; workspaces (tenants) own only their users and memberships.\nHealth probes are public.\n",
    "title": "YouTube Trending Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/videos": {
      "get": {
        "description": "'Latest snapshot' means each video's latest snapshot WITHIN THE FILTERED DATASET (the rows matching the date range, countries and categories), not its globally latest snapshot. Ordering is deterministic: NULLs last, ties broken by video_id.",
        "operationId": "list_videos_api_v1_videos_get",
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VideoSort",
              "default": "views"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Case-insensitive substring of the title or channel title, or an exact video_id; 2-100 characters; wildcards are matched literally.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring of the title or channel title, or an exact video_id; 2-100 characters; wildcards are matched literally.",
              "title": "Search"
            }
          },
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_VideoPage_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Paginated list of trending videos",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/v1/videos/{video_id}": {
      "get": {
        "operationId": "video_detail_api_v1_videos__video_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_VideoDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown video_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Video detail (globally latest snapshot; not filtered)",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/v1/videos/{video_id}/history": {
      "get": {
        "operationId": "video_history_api_v1_videos__video_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_HistoryPoint__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown video_id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Full trending history (all countries and dates; not filtered)",
        "tags": [
          "videos"
        ]
      }
    },
...
  }
}